+++ src/onegov.core/onegov/core/tests/test_layout.py
@@ -22,3 +22,14 @@ def test_format_date():
     assert layout.format_date(dt.date(), 'date') == '17.06.2015'
     assert layout.format_date(dt, 'time') == '12:00'
     assert layout.format_date(dt.time(), 'time') == '12:00'
+
+
+def test_format_number():
+
+    layout = Layout(model=object(), request=object())
+
+    assert layout.format_number(100) == "100"
+    assert layout.format_number(100, 2) == "100.00"
+    assert layout.format_number(1000) == "1'000"
+    assert layout.format_number(1000.00) == "1'000.00"
+    assert layout.format_number(1000.00, 0) == "1'000"